add owner to restaurant and Resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishesForRestaurant/DeleteDishesForRestaurantCommandHandler.cs
@@ -3,10 +3,14 @@ using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.IRepositories;
+using Restaurants.Infrastructure.Authorization.Services;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishesForRestaurant;
 
-public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> _logger, IRestaurantsRepository _restaurantsRepository, IDishesRepository _dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> _logger,
+    IRestaurantsRepository _restaurantsRepository, IDishesRepository _dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService)
+    : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -14,6 +18,9 @@ public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRest
 
         var restaurant = await _restaurantsRepository.GetRestaurantById(request.RestaurantId);
         if (restaurant == null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, Domain.ResourceOperation.Delete))
+            throw new ForbidException();
 
 
         await _dishesRepository.DeleteDishesForRestaurant(restaurant.Dishes);
